+++ RoboRally/src/Main.java
@@ -5,11 +5,10 @@ import javafx.scene.image.Image;
 import javafx.scene.layout.GridPane;
 import javafx.stage.Stage;
 import javafx.fxml.FXMLLoader;
-import modelserver.Server;
+import server.Server;
 import java.io.*;
 import java.util.logging.Logger;
 import utils.Parameter;
-import viewmodel.MapController;
 
 
 /**
@@ -75,8 +74,8 @@ public class Main extends Application {
             try {
                 Stage primaryStage = new Stage();
                 FXMLLoader loader = new FXMLLoader();
-                // Load view for primaryStage
-                loader.setLocation(Main.class.getResource("view/Stage.fxml"));
+                // Load views for primaryStage
+                loader.setLocation(Main.class.getResource("views/Stage.fxml"));
 
                 GridPane stageView;