+++ RoboRally/src/server/game/DamageCards/Worm.java
@@ -1,0 +1,25 @@
+package server.game.DamageCards;
+import static utils.Parameter.WORM_CARDS_AMOUNT;
+
+/**
+ * This class defines a kind of damage.
+ *
+ * @author Vincent Tafferner
+ */
+public class Worm extends DamageCard {
+
+    public Worm() {
+        cardName = "Worm";
+    }
+
+    int cardAmount = WORM_CARDS_AMOUNT;
+
+    /**
+     * This is the method that activates the effect of a damage card. <br>
+     * In this case the robot has to reboot.
+     */
+    @Override
+    public void activateDamage() {
+        // TODO
+    }
+}