+++ RoboRally/src/server/game/GamePhases/UpgradePhase.java
@@ -1,0 +1,11 @@
+package server.game.GamePhases;
+
+/**
+ * This class implements the upgrade phase.
+ * @author Vincent Tafferner
+ * @author Jessica Gerlach
+ */
+public class UpgradePhase {
+
+    //TODO
+}