+++ RoboRally/src/server/game/ProgrammingCards/BackUp.java
@@ -1,0 +1,22 @@
+package server.game.ProgrammingCards;
+
+/**
+ * This class implements the Backup card.
+ *
+ * @author Vincent Tafferner
+ */
+public class BackUp extends ProgrammingCard {
+
+    public BackUp() {
+        cardName = "BackUp";
+    }
+
+    /**
+     * This move will make the robot take one step back. <br>
+     * It is important, that the robot still faces the same direction.
+     */
+    @Override
+    public void moveRobot() {
+        // TODO
+    }
+}